Rename user slice
