+++ src/components/app/index.tsx
@@ -14,7 +14,7 @@ function App(): JSX.Element {
   const globalOffersLoading = useAppSelector(
     (state) => state.globalOffersReducer.loading
   );
-  const userLoading = useAppSelector((state) => state.userReducer.userLoading);
+  const userLoading = useAppSelector((state) => state.userDataReducer.userLoading);
   useAppInit();
 
   if (globalOffersLoading || userLoading) {
