+++ src/store/actions.ts
@@ -6,4 +6,4 @@ export {
 } from './city-offers/actions';
 export { fetchGlobalOffers } from './global-offers/actions';
 export { fetchOffer, postComment } from './offer/actions';
-export { validateUser, login, logout } from './user/actions';
+export { validateUser, login, logout } from './user-data/actions';
